+++ spec/helpers/asserted_environments_helper_spec.rb
@@ -12,10 +12,6 @@ describe AssertedEnvironmentsHelper, type: :helper do
     expect(helper.label_for_asserted_environment(asserted_environment)).to eq('temperate forest biome')
   end
 
-  specify '#label_for_asserted_environment with nil' do
-    expect(helper.label_for_asserted_environment(nil)).to be_nil
-  end
-
   specify '#asserted_environment_tag' do
     expect(helper.asserted_environment_tag(asserted_environment)).to include('temperate forest biome')
   end